added the web.go file and introducted publish path
